+++ Library/ModernUiCustomizedDisplayLib/CustomizedDisplayLibInternal.c
@@ -4,6 +4,10 @@
 
 Copyright (c) 2013-2018, Intel Corporation. All rights reserved.<BR>
 Copyright (c) 2025, Loongson Technology Corporation Limited. All rights reserved.<BR>
+  Copyright (c) 2026, MarsDoge. All rights reserved.<BR>
+  Author: MarsDoge (Dongyan Qian)
+  Open source: https://github.com/MarsDoge/ModernSetupPkg
+
 SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/