+++ Library/ModernUiRendererLib/ModernUiRendererLib.c
@@ -1,6 +1,10 @@
 /** @file
   GOP-backed renderer library for ModernSetupPkg.
 
+  Copyright (c) 2026, MarsDoge. All rights reserved.<BR>
+  Author: MarsDoge (Dongyan Qian)
+  Open source: https://github.com/MarsDoge/ModernSetupPkg
+
   SPDX-License-Identifier: BSD-2-Clause-Patent
 **/