+++ Universal/ModernDisplayEngineDxe/Popup.c
@@ -2,6 +2,10 @@
 Implementation for Hii Popup Protocol.
 
 Copyright (c) 2017, Intel Corporation. All rights reserved.<BR>
+  Copyright (c) 2026, MarsDoge. All rights reserved.<BR>
+  Author: MarsDoge (Dongyan Qian)
+  Open source: https://github.com/MarsDoge/ModernSetupPkg
+
 SPDX-License-Identifier: BSD-2-Clause-Patent
 
 **/